Refactor openfaas crd generation
This commit refactor openfaas crd schema puts it inside a package with
a version name so that it's easy to add new version for openfaas CRD
